Gfpgan download fix (#655)
* restore ability to save files to directories named after prompt

* don't download gfpgan model file if it exists

- Fixes #647

+++ scripts/preload_models.py
@@ -89,8 +89,9 @@ if gfpgan:
         import urllib.request
         model_path = 'https://github.com/TencentARC/GFPGAN/releases/download/v1.3.0/GFPGANv1.3.pth'
         model_dest = 'src/gfpgan/experiments/pretrained_models/GFPGANv1.3.pth'
-        print('downloading gfpgan model file...')
-        urllib.request.urlretrieve(model_path,model_dest)
+        if not os.path.exists(model_dest):
+            print('downloading gfpgan model file...')
+            urllib.request.urlretrieve(model_path,model_dest)
     except Exception:
         import traceback
         print('Error loading GFPGAN:')
